extracted ArrayList from CypherList
