+++ src/Types/CypherList.php
@@ -28,229 +28,12 @@ use function usort;
  *
  * @template TValue
  *
- * @extends AbstractCypherSequence<int, TValue>
+ * @extends ArrayList<TValue>
  *
  * @psalm-immutable
  */
-final class CypherList extends AbstractCypherSequence
+final class CypherList extends ArrayList
 {
-    /**
-     * @param iterable<TValue> $iterable
-     */
-    public function __construct(iterable $iterable = [])
-    {
-        if ($iterable instanceof self) {
-            /** @psalm-suppress InvalidPropertyAssignmentValue */
-            $this->sequence = $iterable->sequence;
-        } else {
-            $this->sequence = [];
-            foreach ($iterable as $value) {
-                $this->sequence[] = $value;
-            }
-        }
-    }
-
-    /**
-     * Returns the first element in the sequence.
-     *
-     * @return TValue
-     */
-    public function first()
-    {
-        if (!array_key_exists(0, $this->sequence)) {
-            throw new OutOfBoundsException('Cannot grab first element of an empty list');
-        }
-
-        return $this->sequence[0];
-    }
-
-    /**
-     * Returns the last element in the sequence.
-     *
-     * @return TValue
-     */
-    public function last()
-    {
-        $key = array_key_last($this->sequence);
-        if (!is_int($key)) {
-            throw new OutOfBoundsException('Cannot grab last element of an empty list');
-        }
-
-        return $this->sequence[$key];
-    }
-
-    /**
-     * @param iterable<TValue> $values
-     *
-     * @return CypherList<TValue>
-     */
-    public function merge($values): CypherList
-    {
-        $tbr = $this->sequence;
-        foreach ($values as $value) {
-            $tbr[] = $value;
-        }
-
-        return new CypherList($tbr);
-    }
-
-    /**
-     * @return CypherList<TValue>
-     */
-    public function reversed(): CypherList
-    {
-        return new CypherList(array_reverse($this->sequence));
-    }
-
-    /**
-     * @return CypherList<TValue>
-     */
-    public function slice(int $offset, int $length = null): CypherList
-    {
-        return new CypherList(array_slice($this->sequence, $offset, $length));
-    }
-
-    /**
-     * @param (pure-callable(TValue, TValue):int)|null $comparator
-     *
-     * @return CypherList<TValue>
-     */
-    public function sorted(callable $comparator = null): CypherList
-    {
-        $tbr = $this->sequence;
-        if ($comparator === null) {
-            sort($tbr);
-        } else {
-            usort($tbr, $comparator);
-        }
-
-        return new CypherList($tbr);
-    }
-
-    /**
-     * @pure
-     */
-    public static function fromIterable(iterable $iterable): AbstractCypherSequence
-    {
-        return new self($iterable);
-    }
-
-    /**
-     * Gets the nth element in the list.
-     *
-     * @throws OutOfBoundsException
-     *
-     * @return TValue
-     */
-    public function get(int $key)
-    {
-        if (!array_key_exists($key, $this->sequence)) {
-            throw new OutOfBoundsException(sprintf('Cannot get item in sequence at position: %s', $key));
-        }
-
-        return $this->sequence[$key];
-    }
-
-    public function getAsString(int $key): string
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toString($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'string');
-        }
-
-        return $tbr;
-    }
-
-    public function getAsInt(int $key): int
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toInt($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'int');
-        }
-
-        return $tbr;
-    }
-
-    public function getAsFloat(int $key): float
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toFloat($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'float');
-        }
-
-        return $tbr;
-    }
-
-    public function getAsBool(int $key): bool
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toBool($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, 'bool');
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @return null
-     */
-    public function getAsNull(int $key)
-    {
-        $this->get($key);
-
-        return TypeCaster::toNull();
-    }
-
-    /**
-     * @template U
-     *
-     * @param class-string<U> $class
-     *
-     * @return U
-     */
-    public function getAsObject(int $key, string $class): object
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toClass($value, $class);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, $class);
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @return CypherMap<mixed>
-     */
-    public function getAsCypherMap(int $key): CypherMap
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toCypherMap($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, CypherMap::class);
-        }
-
-        return $tbr;
-    }
-
-    /**
-     * @return CypherList<mixed>
-     */
-    public function getAsCypherList(int $key): CypherList
-    {
-        $value = $this->get($key);
-        $tbr = TypeCaster::toCypherList($value);
-        if ($tbr === null) {
-            throw new RuntimeTypeException($value, CypherList::class);
-        }
-
-        return $tbr;
-    }
-
     public function getAsDate(int $key): Date
     {
         return $this->getAsObject($key, Date::class);
